requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cloudinary==1.23.0
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-bootstrap4==2.3.1
django-cors-headers==3.5.0
django-rest-auth==0.9.5
django-tinymce==3.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.26.2
